+++ tests/stCactusGraphsTest.c
@@ -15,7 +15,7 @@
 #include <stdlib.h>
 
 static stCactusGraph *g;
-static int32_t nO1 = 1, nO2 = 2, nO3 = 3, nO4 = 4, nO5 = 5, nO6 = 6, nO7 = 7, nO8 = 8, nO9 = 9;
+static int64_t nO1 = 1, nO2 = 2, nO3 = 3, nO4 = 4, nO5 = 5, nO6 = 6, nO7 = 7, nO8 = 8, nO9 = 9;
 static stCactusNode *n1, *n2, *n3, *n4, *n5, *n6, *n7, *n8, *n9;
 static stCactusEdgeEnd *e12, *e21, *e23, *e32, *e13, *e31, *e24, *e42, *e11, *e11r, *e45, *e54, *e56, *e65, *e56r, *e65r, *e58, *e85, *e49,
         *e94, *e37, *e73, *e37r, *e73r;
@@ -28,8 +28,8 @@ static void teardown() {
 }
 
 static void *mergeNodeObjects(void *a, void *b) {
-    int32_t i = *((int32_t *) a);
-    int32_t j = *((int32_t *) b);
+    int64_t i = *((int64_t *) a);
+    int64_t j = *((int64_t *) b);
     return i > j ? b : a;
 }
 
@@ -359,7 +359,7 @@ static void chainStructureTests(CuTest *testCase) {
     invariantChainStructureTests(testCase);
     //Check the chain structure of the bridge ends is uninitialised
     stCactusEdgeEnd *edgeEnds[] = { e24, e42, e45, e49, e54, e58, e85, e94 };
-    for (int32_t i = 0; i < 8; i++) {
+    for (int64_t i = 0; i < 8; i++) {
         CuAssertPtrEquals(testCase, NULL, stCactusEdgeEnd_getLink(edgeEnds[i]));
         CuAssertIntEquals(testCase, 0, stCactusEdgeEnd_getLinkOrientation(edgeEnds[i]));
         CuAssertIntEquals(testCase, 0, stCactusEdgeEnd_isChainEnd(edgeEnds[i]));
@@ -384,7 +384,7 @@ static void testStCactusGraph_unmarkAndMarkCycles(CuTest *testCase) {
     stCactusEdgeEnd *edgeEnds[] = { e12, e13, e11, e11r, e21, e23, e24, e32, e31, e37, e37r, e42, e45, e49, e54, e56, e56r, e58, e65, e65r,
             e73, e73r, e85, e94 };
 
-    for (int32_t i = 0; i < 24; i++) {
+    for (int64_t i = 0; i < 24; i++) {
         CuAssertPtrEquals(testCase, NULL, stCactusEdgeEnd_getLink(edgeEnds[i]));
         CuAssertIntEquals(testCase, 0, stCactusEdgeEnd_getLinkOrientation(edgeEnds[i]));
         CuAssertIntEquals(testCase, 0, stCactusEdgeEnd_isChainEnd(edgeEnds[i]));
@@ -480,16 +480,16 @@ static void testStCactusGraph_randomTest(CuTest *testCase) {
     //return;
     //Creates a problem instances, then checks graph is okay by checking every edge
     //is properly connected, with right number of nodes and that everyone is in a chain
-    for (int32_t test = 0; test < 1000; test++) {
-        int32_t nodeNumber = st_randomInt(0, 100); //1000);
-        int32_t edgeNumber = nodeNumber > 0 ? st_randomInt(0, 200) : 0; //1000) : 0;
-        int32_t longChain = st_randomInt(2, 10);
-        int32_t chainLengthForBigFlower = st_randomInt(0, 20);
-        st_logInfo("We have %i edges and %i nodes in random test %i\n", edgeNumber, nodeNumber, test);
+    for (int64_t test = 0; test < 1000; test++) {
+        int64_t nodeNumber = st_randomInt(0, 100); //1000);
+        int64_t edgeNumber = nodeNumber > 0 ? st_randomInt(0, 200) : 0; //1000) : 0;
+        int64_t longChain = st_randomInt(2, 10);
+        int64_t chainLengthForBigFlower = st_randomInt(0, 20);
+        st_logInfo("We have %" PRIi64 " edges and %" PRIi64 " nodes in random test %" PRIi64 "\n", edgeNumber, nodeNumber, test);
         stCactusGraph *g2 = stCactusGraph_construct();
         stList *nodeObjects = stList_construct3(0, free);
-        for (int32_t i = 0; i < nodeNumber; i++) {
-            int32_t *j = st_malloc(sizeof(int32_t));
+        for (int64_t i = 0; i < nodeNumber; i++) {
+            int64_t *j = st_malloc(sizeof(int64_t));
             j[0] = i;
             stCactusNode_construct(g2, j);
             stList_append(nodeObjects, j);
@@ -498,7 +498,7 @@ static void testStCactusGraph_randomTest(CuTest *testCase) {
         if (nodeNumber > 0) {
             stList *includedNodeObjects = stList_construct(); //Edge construction ensures there is just one component containing edges.
             stList_append(includedNodeObjects, st_randomChoice(nodeObjects));
-            for (int32_t i = 0; i < edgeNumber; i++) {
+            for (int64_t i = 0; i < edgeNumber; i++) {
                 void *nodeObject1 = st_randomChoice(includedNodeObjects);
                 void *nodeObject2 = st_randomChoice(nodeObjects);
                 if (!stList_contains(includedNodeObjects, nodeObject2)) {
@@ -529,7 +529,7 @@ static void testStCactusGraph_randomTest(CuTest *testCase) {
         }
         stCactusGraphNodeIterator_destruct(nodeIt);
         nodeIt = stCactusGraphNodeIterator_construct(g2);
-        int32_t edgeEndNumber = 0;
+        int64_t edgeEndNumber = 0;
         while ((node = stCactusGraphNodeIterator_getNext(nodeIt)) != NULL) {
             CuAssertPtrEquals(testCase, node, stSortedSet_search(nodesInFinalGraph, node));
             stCactusNodeEdgeEndIt edgeEndIt = stCactusNode_getEdgeEndIt(node);
@@ -603,7 +603,7 @@ static void testStCactusGraph_randomTest(CuTest *testCase) {
                 stCactusEdgeEnd *edgeEnd;
                 while ((edgeEnd = stCactusNodeEdgeEndIt_getNext(&edgeIt)) != NULL) {
                     if (stCactusEdgeEnd_isChainEnd(edgeEnd)) {
-                        //st_uglyf("%i %i %i %i %i\n", chainLengthForBigFlower, longChain, stCactusNode_getTotalEdgeLengthOfFlower(node), stCactusNode_getChainNumber(node), stCactusEdgeEnd_getChainLength(edgeEnd));
+                        //st_uglyf("%" PRIi64 " %" PRIi64 " %" PRIi64 " %" PRIi64 " %" PRIi64 "\n", chainLengthForBigFlower, longChain, stCactusNode_getTotalEdgeLengthOfFlower(node), stCactusNode_getChainNumber(node), stCactusEdgeEnd_getChainLength(edgeEnd));
                         CuAssertTrue(testCase, stCactusEdgeEnd_getChainLength(edgeEnd) <= longChain);
                     }
                 }
